refactor: long 타입을 BigInteger로 변경하여 임의 정밀도 지원
- Calculator, NumberConverter, InputValidator, StringCalculator의 타입 변경
- long의 오버플로우 문제 해결
- 임의 크기의 정수 연산 지원
- 모든 테스트 케이스 BigInteger로 수정
- 큰 숫자 테스트 케이스 추가

+++ src/main/java/calculator/Application.java
@@ -1,6 +1,7 @@
 package calculator;
 
 import camp.nextstep.edu.missionutils.Console;
+import java.math.BigInteger;
 
 public class Application {
 
@@ -9,7 +10,7 @@ public class Application {
         String input = Console.readLine();
 
         StringCalculator calculator = new StringCalculator();
-        long result = calculator.calculate(input);
+        BigInteger result = calculator.calculate(input);
 
         System.out.println("결과 : " + result);
     }
